+++ CHANGELOG.md
@@ -61,7 +61,10 @@ list only contains the Pylint checkers enabled by default in PythonTA.
 - Updated dependencies for GitHub Actions to use the latest versions
 - Updated dependabot configuration to auto-update dependencies for GitHub Actions in the future
 - Updated usage messages in `examples/sample_usage/` of `draw_cfg.py` and `print_ast.py` to be accurate on all operating systems
-- Added `z3_vars` attribute to `ControlFlowGraph` to store variables to be used in Z3 solver
+- Added protected `_z3_vars` attribute to `ControlFlowGraph` to store variables to be used in Z3 solver
+- Removed unused imports from `python_ta/cfg/graph.py`
+- Extended functionality of `ExprWrapper` class to include function definitions
+- Added to `CFGVisitor.visit_functiondef` method to update the `ControlFlowGraph`'s `_z3_vars` attribute
 
 ## [2.7.0] - 2024-12-14